Make new file, add merge goal task
