+++ optd-core/src/optimizer/memo.rs
@@ -45,7 +45,11 @@ pub struct MergeGroupResult {
 pub struct MergeGoalResult {
     /// Goals that were merged along with their potential best costed expression.
     /// All expressions listed in `merged_exprs` must be contained within these goals.
-    pub merged_goals: Vec<(GoalId, Option<(PhysicalExpressionId, Cost)>)>,
+    pub merged_goals: Vec<(
+        GoalId,
+        Option<(PhysicalExpressionId, Cost)>,
+        Vec<PhysicalExpressionId>,
+    )>,
 
     /// ID of the new representative goal id.
     pub new_repr_goal_id: GoalId,
